refactor #25 : feign client 컨벤션 맞춰 수정
