--- src/main/java/com/irum/orderservice/openfeign/config/FeignConfig.java
@@ -1,14 +1,0 @@
-package com.irum.orderservice.openfeign.config;
-
-import feign.Logger;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class FeignConfig {
-
-    @Bean
-    Logger.Level feignLoggerLevel() {
-        return Logger.Level.FULL;
-    }
-}
