+++ src/main/java/com/irum/orderservice/openfeign/payment/dto/request/CreatePaymentRequest.java
@@ -1,6 +1,6 @@
 package com.irum.orderservice.openfeign.payment.dto.request;
 
-import com.irum.orderservice.openfeign.payment.dto.emuns.PaymentCorp;
+import com.irum.orderservice.openfeign.payment.emuns.PaymentCorp;
 import lombok.Builder;
 
 @Builder
