+++ src/main/java/com/irum/orderservice/openfeign/payment/dto/response/PaymentResponse.java
@@ -1,6 +1,6 @@
 package com.irum.orderservice.openfeign.payment.dto.response;
 
-import com.irum.orderservice.openfeign.payment.dto.emuns.PaymentMethod;
-import com.irum.orderservice.openfeign.payment.dto.emuns.PaymentStatus;
+import com.irum.orderservice.openfeign.payment.emuns.PaymentMethod;
+import com.irum.orderservice.openfeign.payment.emuns.PaymentStatus;
 
 public record PaymentResponse(PaymentStatus paymentStatus, PaymentMethod paymentMethod) {}
