+++ src/main/java/com/irum/orderservice/openfeign/product/client/ProductClient.java
@@ -1,25 +1,31 @@
-package com.irum.orderservice.openfeign.product.api;
+package com.irum.orderservice.openfeign.product.client;
 
 import com.irum.orderservice.openfeign.config.FeignConfig;
 import com.irum.orderservice.openfeign.product.dto.request.ProductInternalRequest;
 import com.irum.orderservice.openfeign.product.dto.request.RollbackStockRequest;
 import com.irum.orderservice.openfeign.product.dto.response.ProductInternalResponse;
+import com.irum.orderservice.openfeign.product.dto.response.StoreResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
+
+import java.util.UUID;
 
 @FeignClient(
         name = "PRODUCT-SERVICE",
-        url = "product-service/internal/products",
-        configuration = FeignConfig.class)
-public interface ProductAPI {
+        url = "/internal")
+public interface ProductClient {
 
     /** 재고 롤백 */
-    @PatchMapping("/rollback")
+    @PatchMapping("/products/rollback")
     void rollbackStock(@RequestBody RollbackStockRequest request);
 
     /** 주문 - 재고 차감 */
-    @GetMapping("/stock")
+    @GetMapping("/products/stock")
     ProductInternalResponse updateStock(@RequestBody ProductInternalRequest request);
+
+    @GetMapping("/stores/{storeId}/owner")
+    StoreResponse getStoreId(@PathVariable UUID storeId);
 }
